--- a/transformations/url_decode_uni.go
+++ b/transformations/url_decode_uni.go
@@ -128,5 +128,5 @@ func inplaceUniDecode(input string, d []byte, pos int) string {
 		}
 	}
 
-	return string(d[0:c])
+	return strings.WrapUnsafe(d[0:c])
 }