--- a/transformations/url_decode_uni_test.go
+++ b/transformations/url_decode_uni_test.go
@@ -22,7 +22,9 @@ func BenchmarkURLDecode(b *testing.B) {
 			tt := tc
 			b.Run(fmt.Sprintf("%s/%s", mode, tt), func(b *testing.B) {
 				for i := 0; i < b.N; i++ {
-					f(tt)
+					if _, err := f(tt); err != nil {
+						b.Fatal(err)
+					}
 				}
 			})
 		}