Transformation cache will help a lot but anyways we shouldn't allocate where there's no need to

Also wraps the result buffer instead of copying

After
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/transformations
BenchmarkURLDecode
BenchmarkURLDecode/normal/
BenchmarkURLDecode/normal/-10         	563965651	         2.112 ns/op
BenchmarkURLDecode/normal/helloworld
BenchmarkURLDecode/normal/helloworld-10         	212564259	         5.637 ns/op
BenchmarkURLDecode/normal/hello+world
BenchmarkURLDecode/normal/hello+world-10        	46705327	        25.33 ns/op
BenchmarkURLDecode/normal/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89
BenchmarkURLDecode/normal/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89-10         	14283956	        84.26 ns/op
BenchmarkURLDecode/unicode/
BenchmarkURLDecode/unicode/-10                                                                       	568736440	         2.114 ns/op
BenchmarkURLDecode/unicode/helloworld
BenchmarkURLDecode/unicode/helloworld-10                                                             	213121411	         5.636 ns/op
BenchmarkURLDecode/unicode/hello+world
BenchmarkURLDecode/unicode/hello+world-10                                                            	46748011	        25.32 ns/op
BenchmarkURLDecode/unicode/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89
BenchmarkURLDecode/unicode/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89-10        	12273687	        97.73 ns/op
PASS
```

No WrapUnsafe
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/transformations
BenchmarkURLDecode
BenchmarkURLDecode/normal/
BenchmarkURLDecode/normal/-10         	219029828	         5.482 ns/op
BenchmarkURLDecode/normal/helloworld
BenchmarkURLDecode/normal/helloworld-10         	45085591	        25.96 ns/op
BenchmarkURLDecode/normal/hello+world
BenchmarkURLDecode/normal/hello+world-10        	42104770	        27.85 ns/op
BenchmarkURLDecode/normal/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89
BenchmarkURLDecode/normal/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89-10         	12552880	        95.92 ns/op
BenchmarkURLDecode/unicode/
BenchmarkURLDecode/unicode/-10                                                                       	219463183	         5.472 ns/op
BenchmarkURLDecode/unicode/helloworld
BenchmarkURLDecode/unicode/helloworld-10                                                             	46115858	        26.04 ns/op
BenchmarkURLDecode/unicode/hello+world
BenchmarkURLDecode/unicode/hello+world-10                                                            	43190130	        27.85 ns/op
BenchmarkURLDecode/unicode/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89
BenchmarkURLDecode/unicode/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89-10        	10971687	       109.3 ns/op
PASS
```

Before
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/transformations
BenchmarkURLDecode
BenchmarkURLDecode/normal/
BenchmarkURLDecode/normal/-10         	568033779	         2.110 ns/op
BenchmarkURLDecode/normal/helloworld
BenchmarkURLDecode/normal/helloworld-10         	212907738	         5.631 ns/op
BenchmarkURLDecode/normal/hello+world
BenchmarkURLDecode/normal/hello+world-10        	45469758	        25.55 ns/op
BenchmarkURLDecode/normal/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89
BenchmarkURLDecode/normal/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89-10         	12481731	        96.75 ns/op
BenchmarkURLDecode/unicode/
BenchmarkURLDecode/unicode/-10                                                                       	568933278	         2.110 ns/op
BenchmarkURLDecode/unicode/helloworld
BenchmarkURLDecode/unicode/helloworld-10                                                             	213502885	         5.636 ns/op
BenchmarkURLDecode/unicode/hello+world
BenchmarkURLDecode/unicode/hello+world-10                                                            	46644056	        25.55 ns/op
BenchmarkURLDecode/unicode/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89
BenchmarkURLDecode/unicode/%E3%83%8F%E3%83%AD%E3%83%BC%E3%83%AF%E3%83%BC%E3%83%AB%E3%83%89-10        	10664019	       112.8 ns/op
PASS
```